+++ routes/bots/endpoints/get_bot/route.go
@@ -272,7 +272,7 @@ func Route(d uapi.RouteData, r *http.Request) uapi.HttpResponse {
 	}
 
 	bot.User = botUser
-	botassets.ApplySelfStatus(bot.User, bot.SelfStatus.String)
+	botassets.ApplySelfStatus(bot.User, bot.SelfStatus.String, bot.Servers, bot.LastStatsPost)
 
 	var uniqueClicks int64
 	err = state.Pool.QueryRow(d.Context, "SELECT cardinality(unique_clicks) AS unique_clicks FROM bots WHERE bot_id = $1", bot.BotID).Scan(&uniqueClicks)